requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
boto3==1.26.155
botocore==1.29.155
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.1
cryptography==41.0.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.6
Django==4.2.2
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-environ==0.10.0
django-import-export==3.2.0
django-js-asset==2.0.0
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
et-xmlfile==1.1.0
filelock==3.12.2
Flask==2.0.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==1.0.1
Markdown==3.4.3
MarkupPy==1.14
MarkupSafe==2.0.1
mutagen==1.46.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.5.0
platformdirs==3.6.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stripe==3.0.0
tablib==3.5.0
toml==0.10.2
typing_extensions==4.6.3
urllib3==1.26.16
virtualenv==20.23.1
Werkzeug==2.0.1
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
